+++ examples/sample.js
@@ -1,0 +1,14 @@
+
+
+
+function brandon() {
+  console.log(arguments);
+}
+
+
+function anna() {
+  brandon.apply(this, arguments);
+}
+
+
+anna('You', 'are', 'sweet')